requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==41.0.7
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==3.14.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.0
sqlparse==0.5.4
typing_extensions==4.15.0
